# docs/swagger.yaml
basePath: /
definitions:
  handler.ResponseSuccess:
    properties:
      data: {}
      message:
        type: string
    type: object
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
  handler.errorResponseData:
    properties:
      data: {}
      message:
        type: string
    type: object
  model.Admin:
    properties:
      password:
        default: "12345"
        type: string
      user_name:
        default: justice
        type: string
    type: object
  model.Contact:
    properties:
      firstName:
        type: string
      lastName:
        type: string
      phoneNumber:
        type: string
      text:
        type: string
      typeService:
        type: string
    type: object
  model.ContactFull:
    properties:
      created_at:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      phone_number:
        type: string
      text:
        type: string
      type_service:
        type: string
    type: object
...
  model.HomeFull:
    properties:
      id:
        type: integer
      post_body:
        type: string
      post_body_ru:
        type: string
      post_date:
        type: string
      post_img_path:
        type: string
      post_img_url:
        type: string
      post_title:
        type: string
      post_title_ru:
        type: string
    type: object
  model.HomePost:
    properties:
      post_body:
        type: string
      post_body_ru:
        type: string
      post_img_url:
        type: string
      post_title:
        type: string
      post_title_ru:
        type: string
    type: object
...
  model.allContacts:
    properties:
      allContact:
        items:
          $ref: '#/definitions/model.ContactFull'
        type: array
    type: object
  model.allHome:
    properties:
      allHome:
        items:
          $ref: '#/definitions/model.HomeFull'
        type: array
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API Server for Norbekov App
  title: Norbekov API
  version: "1.0"
paths:
  /admin/create:
    post:
      consumes:
      - application/json
      description: create admin_post
      operationId: create-admin_post
      parameters:
      - description: home info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.Admin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Create Admin
      tags:
      - Admin
  /admin/delete:
    delete:
      consumes:
      - application/json
      description: delete admin
      operationId: delete-admin
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get Delete Admin
      tags:
      - Admin
  /admin/login:
    post:
      consumes:
      - application/json
      description: login admin
      operationId: login-admin
      parameters:
      - description: admin info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.Admin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Login Admin
      tags:
      - Admin
  /api/home/create:
    post:
      consumes:
      - application/json
      description: create home_post
      operationId: create-home_post
      parameters:
      - description: home info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.HomePost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create HomePost
      tags:
      - Home
  /api/home/delete:
    delete:
      consumes:
      - application/json
      description: delete post
      operationId: delete-posts
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.HomeFull'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Delete Post
      tags:
      - Home
  /api/home/update/{id}:
    put:
      consumes:
      - application/json
      description: Update home by id
      operationId: update-home-id
      parameters:
      - description: Param ID
        in: path
        name: id
        required: true
        type: integer
      - description: home info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.HomePost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update  Home By ID
      tags:
      - Home
  /api/home/upload-img/{id}:
    patch:
      consumes:
      - application/json
      - multipart/form-data
      description: Upload Home Image
      operationId: upload-image
      parameters:
      - description: Param ID
        in: path
        name: id
        required: true
        type: integer
      - description: file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      - application/octet-stream
      - image/png
      - image/jpeg
      - image/jpg
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Upload Home Image
      tags:
      - Home
...
  /contact/create:
    post:
      consumes:
      - application/json
      description: create contact_post
      operationId: create-contact_post
      parameters:
      - description: home info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.Contact'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Create Contact
      tags:
      - Contact
  /contact/get:
    get:
      consumes:
      - application/json
      description: get  contacts
      operationId: get-contacts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.allContacts'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get contacts
      tags:
      - Contact
  /home/get:
    get:
      consumes:
      - application/json
      description: get home post
      operationId: get-home
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.HomeFull'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get home
      tags:
      - Home
  /homes/get:
    get:
      consumes:
      - application/json
      description: get  homes
      operationId: get-homes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.allHome'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get homes
      tags:
      - Home
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"